requirements.txt
langchain==0.3.1
langchain_community==0.3.1
langchain_ollama==0.2.0
langchain_qdrant==0.1.4
ollama==0.3.3
pdfplumber==0.11.4
python-dotenv==1.0.1
qdrant_client==1.11.3
streamlit==1.32.2
